Moved sub-packages configs to their owners, out of a central config
